requirement.txt
discord.py-self
halo
tkinter
python-telegram-bot
openai
dotenv